# Add service label on metrics