@davidor 2019-04-30
@eloycoto the new metrics should be added to https://github.com/3scale/APIcast/blob/master/doc/prometheus-metrics.md