@eloycoto 2019-04-30
I think that I found the issue with the test[0]. 

On the executor, where the phase `log` happens, the ngx.ctx.context is nil, so
it creates a new context and the old one is not updated.

https://github.com/3scale/APIcast/blob/66168b7a15655f0ea473d862a41ae9fdbb3fa912/gateway/src/apicast/executor.lua#L40-L51

The main problem on log phase, if the batch policy or cache policy is in place
the context.service is not defined, so the Prometheus metrics raise an error
that the service.id cannot be found.

I debug all the code, and the issue is fixed if we add in this the following
code `require('resty.ctx').apply()`, if we add this the code is working
correctly. 

https://github.com/3scale/APIcast/blob/66168b7a15655f0ea473d862a41ae9fdbb3fa912/gateway/conf.d/apicast.conf#L68-L71

I think that it's safe enough, but I want to make sure that I do not introduce
a new issue. Also IMHO we should trigger a error message on
`executor:shared_build_context` that if the context does not exists maybe we
need to notice. (But I do not have the full overview of the project)

[0] https://github.com/3scale/APIcast/blob/master/t/apicast-policy-3scale-batcher.t#L403-L483