@eloycoto 2019-04-30
Added the patch on log phase in apicast.conf, now a different test is failing:

https://github.com/3scale/APIcast/blob/66168b7a15655f0ea473d862a41ae9fdbb3fa912/t/apicast.t#L621-L631

Should a request to invalid service reach `log` phase? Should be denied in the auth phase? @davidor any insight here?