@davidor 2019-05-02
@eloycoto The log phase will run even when the `find_service` policy cannot find a service for the request. In that case `context.service` is going to be nil. That explains why the test `TEST 17: invalid service` that you pasted above fails.

We'll need to take that case into account when updating the Prometheus metrics.