@eloycoto 2019-05-02
@davidor what about this commit? Is that correct?
https://github.com/3scale/APIcast/pull/1024/commits/190962e6449afa53418e9718cc0a4bd58ea5e05b