@davidor 2019-05-02
> @davidor what about this commit? Is that correct?
> [190962e](https://github.com/3scale/APIcast/commit/190962e6449afa53418e9718cc0a4bd58ea5e05b)

@eloycoto and I have been investigating this. Adding `require('resty.ctx').apply()` to the log by lua block in `gateway/conf.d/apicast.conf` is not the solution. When starting APIcast with `bin/apicast` we saw that the context is shared correctly in all the phases, including the `log` one. There's only one test that fails and the reason is that it is the only one that defines its own `rewrite_by_lua`: https://github.com/3scale/APIcast/blob/66168b7a15655f0ea473d862a41ae9fdbb3fa912/t/apicast-policy-3scale-batcher.t#L453

Calling `require('resty.ctx').apply()` in that rewrite block fixes the issue.