@eloycoto 2019-05-03
Hi, 

PR Updated, all the changes in the same commits, the PR is quite small, issues
fixed:

- For the context issues, now in the `M:log` is checking that the
  context.service exists if not a empty service will be used. So the test that
  had failed last Tuesday are now ok.  
- For the service, no longer the key "all", a empty string will be send. 
- For `t/prometheus-metrics.t` a new helper was added, is like
  response_body_like but with multiple options. I did not add this helper in a
  proper library because I think that it's only needed here, but if you think
  that it's good to move it, I'm happy to move it.   

Please, review it again. 
Eloy