@davidor 2019-05-03
@eloycoto Sorry, I thought about this after approving the PR. Maybe we should use the service name or  instead of the ID as the label of the metric.

I'm thinking about how this can be used to show data in something like Grafana. The ID will not give much information to the people operating APIcast. It's likely that they'll need to go to the admin portal to check which service that ID corresponds to. Maybe we can save them that step.