@eloycoto 2019-05-03
For the record, we discussed offline with @davidor:

If we use the ID, the number will never change. 
There is a name that the users can update, but that means that the historical information will be lost. 
There is a system_name variable that will work, but if you update the Name in the UI the system_name will not be updated, so it's like an old id.

We are going to keep the ID that it's the most reasonable one.