@eloycoto 2019-05-06
Hi, 

Made some perfomance testing around the multiple services in the labels, it's
not a big change, I didn't notice any memory issue, is around 200req/sec slower
and adds a 2ms latency: 

Config: https://gist.githubusercontent.com/eloycoto/37b23b7c74a21645e6a4c37f41197c53/raw/447eaf62434e50cd1e0253b622443f0d6de99d64/3scale-100-nodes.json
wrk setup:

```
function request()
   wrk.headers["Host"] = tostring(math.random(0,99))
   return wrk.format(nil, "/?user_key=foo")
end
```

Result with extended metrics:
```
root@ubuntu-s-4vcpu-8gb-ams3-01:~/APIcast# wrk  --connections 100 --threads 10 --duration 300 'http://172.18.0.3:8080/?user_key=foo' -s wrk-code.lua
Running 5m test @ http://172.18.0.3:8080/?user_key=foo
  10 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    27.49ms    7.93ms 141.59ms   72.14%
    Req/Sec   365.66     40.38     0.95k    75.36%
  1092416 requests in 5.00m, 234.36MB read
Requests/sec:   3640.26
Transfer/sec:    799.69KB
```

Upstream status metrics:
```
bash-4.2# curl http://localhost:9421/metrics -s | grep ^upstream_status
upstream_status{status="200",service="0"} 10772
upstream_status{status="200",service="1"} 10933
upstream_status{status="200",service="10"} 10410
upstream_status{status="200",service="11"} 11157
upstream_status{status="200",service="12"} 10465
upstream_status{status="200",service="13"} 11176
upstream_status{status="200",service="14"} 10485
upstream_status{status="200",service="15"} 10869
upstream_status{status="200",service="16"} 11551
upstream_status{status="200",service="17"} 11190
upstream_status{status="200",service="18"} 11373
upstream_status{status="200",service="19"} 10638
upstream_status{status="200",service="2"} 10568
upstream_status{status="200",service="20"} 10732
upstream_status{status="200",service="21"} 11071
upstream_status{status="200",service="22"} 11077
upstream_status{status="200",service="23"} 10724
upstream_status{status="200",service="24"} 10985
upstream_status{status="200",service="25"} 10688
upstream_status{status="200",service="26"} 10419
upstream_status{status="200",service="27"} 10913
upstream_status{status="200",service="28"} 11312
upstream_status{status="200",service="29"} 11288
upstream_status{status="200",service="3"} 11297
upstream_status{status="200",service="30"} 10395
upstream_status{status="200",service="31"} 10977
upstream_status{status="200",service="32"} 10810
upstream_status{status="200",service="33"} 11141
upstream_status{status="200",service="34"} 10920
upstream_status{status="200",service="35"} 11128
upstream_status{status="200",service="36"} 10456
upstream_status{status="200",service="37"} 10756
upstream_status{status="200",service="38"} 11159
upstream_status{status="200",service="39"} 10872
upstream_status{status="200",service="4"} 10773
upstream_status{status="200",service="40"} 10346
upstream_status{status="200",service="41"} 10553
upstream_status{status="200",service="42"} 11502
upstream_status{status="200",service="43"} 10724
upstream_status{status="200",service="44"} 10953
upstream_status{status="200",service="45"} 11275
upstream_status{status="200",service="46"} 10420
upstream_status{status="200",service="47"} 11296
upstream_status{status="200",service="48"} 11134
upstream_status{status="200",service="49"} 10559
upstream_status{status="200",service="5"} 10817
upstream_status{status="200",service="50"} 10998
upstream_status{status="200",service="51"} 10654
upstream_status{status="200",service="52"} 10794
upstream_status{status="200",service="53"} 10395
upstream_status{status="200",service="54"} 11651
upstream_status{status="200",service="55"} 10933
upstream_status{status="200",service="56"} 11145
upstream_status{status="200",service="57"} 10777
upstream_status{status="200",service="58"} 10043
upstream_status{status="200",service="59"} 11505
upstream_status{status="200",service="6"} 10780
upstream_status{status="200",service="60"} 10272
upstream_status{status="200",service="61"} 10658
upstream_status{status="200",service="62"} 10847
upstream_status{status="200",service="63"} 11367
upstream_status{status="200",service="64"} 11336
upstream_status{status="200",service="65"} 10528
upstream_status{status="200",service="66"} 10825
upstream_status{status="200",service="67"} 11279
upstream_status{status="200",service="68"} 11098
upstream_status{status="200",service="69"} 10708
upstream_status{status="200",service="7"} 11377
upstream_status{status="200",service="70"} 11280
upstream_status{status="200",service="71"} 11224
upstream_status{status="200",service="72"} 10878
upstream_status{status="200",service="73"} 10995
upstream_status{status="200",service="74"} 11010
upstream_status{status="200",service="75"} 10953
upstream_status{status="200",service="76"} 11471
upstream_status{status="200",service="77"} 10683
upstream_status{status="200",service="78"} 10551
upstream_status{status="200",service="79"} 10088
upstream_status{status="200",service="8"} 11209
upstream_status{status="200",service="80"} 11152
upstream_status{status="200",service="81"} 11418
upstream_status{status="200",service="82"} 10981
upstream_status{status="200",service="83"} 10945
upstream_status{status="200",service="84"} 10713
upstream_status{status="200",service="85"} 10293
upstream_status{status="200",service="86"} 11375
upstream_status{status="200",service="87"} 10748
upstream_status{status="200",service="88"} 10630
upstream_status{status="200",service="89"} 11301
upstream_status{status="200",service="9"} 11493
upstream_status{status="200",service="90"} 11578
upstream_status{status="200",service="91"} 10510
upstream_status{status="200",service="92"} 11049
upstream_status{status="200",service="93"} 11309
upstream_status{status="200",service="94"} 10555
upstream_status{status="200",service="95"} 10859
upstream_status{status="200",service="96"} 11179
upstream_status{status="200",service="97"} 10857
upstream_status{status="200",service="98"} 11392
upstream_status{status="200",service="99"} 10790
bash-4.2#
```

Without extended metrics:

wrk output:
```
root@ubuntu-s-4vcpu-8gb-ams3-01:~/APIcast# wrk  --connections 100 --threads 10 --duration 300 'http://172.18.0.3:8080/?user_key=foo' -s wrk-code.lua
Running 5m test @ http://172.18.0.3:8080/?user_key=foo
  10 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    25.81ms    7.61ms 118.59ms   72.33%
    Req/Sec   389.60     46.55     1.07k    79.60%
  1163878 requests in 5.00m, 249.69MB read
Requests/sec:   3878.50
Transfer/sec:    852.02KB
root@ubuntu-s-4vcpu-8gb-ams3-01:~/APIcast#
```

Metrics:
```
bash-4.2# curl http://localhost:9421/metrics -s | grep ^upstream_status
upstream_status{status="200",service=""} 1163966
bash-4.2#
```